Create LTV_example.csv
Added an example table with LTV calculations for a 30-day user cohort, generated from SQL queries. The data includes income, discounts, marketing costs, and the calculated LTV for each advertising campaign. This file is a result of the code execution and will be used for further analysis and data visualization in Tableau.